+++ resources/views/Index.blade.php
@@ -10,8 +10,8 @@
     <link
         href="https://fonts.googleapis.com/css2?family=Outfit:wght@300;400;500;600;700&family=Space+Mono:wght@400;700&display=swap"
         rel="stylesheet">
-    <link rel="stylesheet"
-        href="{{ asset('css/templatemo-glass-admin-style.css') }}?v={{ filemtime(public_path('css/templatemo-glass-admin-style.css')) }}">
+    <link rel="stylesheet" href="{{ asset('css/shared.css') }}?v={{ filemtime(public_path('css/shared.css')) }}">
+    <link rel="stylesheet" href="{{ asset('css/index.css') }}?v={{ filemtime(public_path('css/index.css')) }}">
     @livewireStyles
 </head>
 
@@ -20,10 +20,9 @@
 
     <x-site-footer />
 
-    <script
-        src="{{ asset('js/templatemo-glass-admin-script.js') }}?v={{ filemtime(public_path('js/templatemo-glass-admin-script.js')) }}"></script>
-    <script
-        src="{{ asset('js/rss-copy-to-clipboard.js') }}?v={{ filemtime(public_path('js/rss-copy-to-clipboard.js')) }}"></script>
+    <script src="{{ asset('js/app-ui.js') }}?v={{ filemtime(public_path('js/app-ui.js')) }}"></script>
+    <script src="{{ asset('js/rss-copy-to-clipboard.js') }}?v={{ filemtime(public_path('js/rss-copy-to-clipboard.js')) }}">
+    </script>
     <script src="{{ asset('js/localize-datetime.js') }}?v={{ filemtime(public_path('js/localize-datetime.js')) }}"></script>
     @livewireScripts
 </body>