+++ resources/views/components/site-footer.blade.php
@@ -1,3 +1,3 @@
-<footer class="site-footer">
+<footer class="app-footer">
     <p>YouTube RSS</p>
 </footer>